feat: fix integration test bug
