+++ user-management/tests/Stickerlandia.UserManagement.IntegrationTest/Drivers/AzureServiceBusMessaging.cs
@@ -8,10 +8,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025 Datadog, Inc.
 
-using System.Text.Json;
+using System.Text;
 using Azure.Messaging.ServiceBus;
-using CloudNative.CloudEvents;
-using CloudNative.CloudEvents.SystemTextJson;
 
 namespace Stickerlandia.UserManagement.IntegrationTest.Drivers;
 
@@ -19,25 +17,11 @@ internal sealed class AzureServiceBusMessaging(string connectionString) : IMessa
 {
     private readonly ServiceBusClient _client = new(connectionString);
 
-    public async Task SendMessageAsync(string queueName, object message)
+    public async Task SendMessageAsync(string queueName, string messageJson)
     {
         var sender = _client.CreateSender(queueName);
 
-        var cloudEvent = new CloudEvent(CloudEventsSpecVersion.V1_0)
-        {
-            Id = Guid.NewGuid().ToString(),
-            Source = new Uri("https://stickerlandia.com"),
-            Type = queueName,
-            Time = DateTime.UtcNow,
-            Data = JsonDocument.Parse(JsonSerializer.Serialize(message)).RootElement
-        };
-        var formatter = new JsonEventFormatter();
-        var encoded = formatter.EncodeStructuredModeMessage(cloudEvent, out _);
-
-        var serviceBusMessage = new ServiceBusMessage(encoded)
-        {
-            ContentType = "application/cloudevents+json"
-        };
+        var serviceBusMessage = new ServiceBusMessage(Encoding.UTF8.GetBytes(messageJson));
 
         await sender.SendMessageAsync(serviceBusMessage);
     }
