+++ user-management/tests/Stickerlandia.UserManagement.IntegrationTest/Drivers/EventBridgeMessaging.cs
@@ -8,12 +8,8 @@
 // This product includes software developed at Datadog (https://www.datadoghq.com/).
 // Copyright 2025 Datadog, Inc.
 
-using System.Text;
-using System.Text.Json;
 using Amazon.EventBridge;
 using Amazon.EventBridge.Model;
-using CloudNative.CloudEvents;
-using CloudNative.CloudEvents.SystemTextJson;
 
 namespace Stickerlandia.UserManagement.IntegrationTest.Drivers;
 
@@ -21,19 +17,8 @@ internal sealed class EventBridgeMessaging(string environment) : IMessaging, IAs
 {
     private readonly AmazonEventBridgeClient _client = new();
 
-    public async Task SendMessageAsync(string queueName, object message)
+    public async Task SendMessageAsync(string queueName, string messageJson)
     {
-        var cloudEvent = new CloudEvent(CloudEventsSpecVersion.V1_0)
-        {
-            Id = Guid.NewGuid().ToString(),
-            Source = new Uri("https://stickerlandia.com"),
-            Type = queueName,
-            Time = DateTime.UtcNow,
-            Data = JsonDocument.Parse(JsonSerializer.Serialize(message)).RootElement
-        };
-        var formatter = new JsonEventFormatter();
-        var encoded = formatter.EncodeStructuredModeMessage(cloudEvent, out _);
-
         await _client.PutEventsAsync(new PutEventsRequest
         {
             Entries = new List<PutEventsRequestEntry>
@@ -43,7 +28,7 @@ internal sealed class EventBridgeMessaging(string environment) : IMessaging, IAs
                     EventBusName = $"Stickerlandia-Shared-{environment}",
                     Source = $"{environment}.stickers",
                     DetailType = queueName,
-                    Detail = Encoding.UTF8.GetString(encoded.Span)
+                    Detail = messageJson
                 }
             }
         });
